+++ .prettierignore
@@ -5,3 +5,4 @@ d2renderers/d2latex/polyfills.js
 d2renderers/d2latex/setup.js
 d2renderers/d2sketch/rough.js
 lib/png/generate_png.js
+d2js
