+++ d2js/js/.prettierignore
@@ -1,0 +1,1 @@
+src/platform.browser.js
