+++ d2js/js/src/platform.browser.js
@@ -1,4 +1,5 @@
 import { wasmBinary, wasmExecJs } from "./wasm-loader.browser.js";
+import workerScript from "./worker.js" with { type: "text" };
 
 export async function loadFile(path) {
   if (path === "./d2.wasm") {
@@ -11,13 +12,6 @@ export async function loadFile(path) {
 }
 
 export async function createWorker() {
-  let response = await fetch(new URL("./worker.js", import.meta.url));
-  if (!response.ok)
-    throw new Error(
-      `Failed to load worker.js: ${response.status} ${response.statusText}`
-    );
-  let workerScript = await response.text();
-
   let blob = new Blob([wasmExecJs, workerScript], {
     type: "text/javascript;charset=utf-8",
   });
